Fix heisenbug with macOS encountering a null block pointer for unknown reason
We suspect it's a null pointer because it reports a memory violation at address equal to the offset of the `invoke` field.
This might be caused by deallocation but unclear why: we hold onto the relevant objects.

This only happens for the logging callbacks, and only from inside the test harness.